analysis: Move idle-related analysis from cpus to idle analysis
