+++ lisa/analysis/cpus.py
@@ -17,12 +17,7 @@
 
 """ CPUs Analysis Module """
 
-import operator
-from functools import reduce
-
 import pandas as pd
-
-from trappy.utils import handle_duplicate_index
 
 from lisa.utils import memoized
 from lisa.analysis.base import AnalysisBase, requires_events
@@ -62,97 +57,6 @@ class CpusAnalysis(AnalysisBase):
         ctx_sw_df.index.name = 'cpu'
 
         return ctx_sw_df
-
-    @requires_events(['cpu_idle'])
-    def df_cpu_wakeups(self, cpus=None):
-        """"
-        Get a DataFrame showing when a CPU was woken from idle
-
-        :param cpus: List of CPUs to find wakeups for. If None, all CPUs.
-        :type cpus: list(int) or None
-
-        :returns: A :class:`pandas.DataFrame` with
-
-          * A ``cpu`` column (the CPU that woke up at the row index)
-        """
-        cpus = cpus or list(range(self._trace.cpus_count))
-
-        sr = pd.Series()
-        for cpu in cpus:
-            cpu_sr = self._trace.getCPUActiveSignal(cpu)
-            cpu_sr = cpu_sr[cpu_sr == 1]
-            cpu_sr = cpu_sr.replace(1, cpu)
-            sr = sr.append(cpu_sr)
-
-        return pd.DataFrame({'cpu': sr}).sort_index()
-
-    @memoized
-    @requires_events(['cpu_idle'])
-    def signal_cpu_active(self, cpu):
-        """
-        Build a square wave representing the active (i.e. non-idle) CPU time
-
-        :param cpu: CPU ID
-        :type cpu: int
-
-        :returns: A :class:`pandas.Series` that equals 1 at timestamps where the
-          CPU is reported to be non-idle, 0 otherwise
-        """
-        idle_df = self._trace.df_events('cpu_idle')
-        cpu_df = idle_df[idle_df.cpu_id == cpu]
-
-        cpu_active = cpu_df.state.apply(
-            lambda s: 1 if s == -1 else 0
-        )
-
-        start_time = 0.0
-        if not self._trace.ftrace.normalized_time:
-            start_time = self._trace.ftrace.basetime
-
-        if cpu_active.empty:
-            cpu_active = pd.Series([0], index=[start_time])
-        elif cpu_active.index[0] != start_time:
-            entry_0 = pd.Series(cpu_active.iloc[0] ^ 1, index=[start_time])
-            cpu_active = pd.concat([entry_0, cpu_active])
-
-        # Fix sequences of wakeup/sleep events reported with the same index
-        return handle_duplicate_index(cpu_active)
-
-    @requires_events(signal_cpu_active.required_events)
-    def signal_cluster_active(self, cluster):
-        """
-        Build a square wave representing the active (i.e. non-idle) cluster time
-
-        :param cluster: list of CPU IDs belonging to a cluster
-        :type cluster: list(int)
-
-        :returns: A :class:`pandas.Series` that equals 1 at timestamps where at
-          least one CPU is reported to be non-idle, 0 otherwise
-        """
-        active = self.signal_cpu_active(cluster[0]).to_frame(name=cluster[0])
-        for cpu in cluster[1:]:
-            active = active.join(
-                self.signal_cpu_active(cpu).to_frame(name=cpu),
-                how='outer'
-            )
-
-        active.fillna(method='ffill', inplace=True)
-        # There might be NaNs in the signal where we got data from some CPUs
-        # before others. That will break the .astype(int) below, so drop rows
-        # with NaN in them.
-        active.dropna(inplace=True)
-
-        # Cluster active is the OR between the actives on each CPU
-        # belonging to that specific cluster
-        cluster_active = reduce(
-            operator.or_,
-            [cpu_active.astype(int) for _, cpu_active in
-             active.items()]
-        )
-
-        return cluster_active
-
-
 
 ###############################################################################
 # Plotting Methods
